style(cargo): satisfy workspace custom lints
